+++ dot/parachain/overseer/overseer.go
@@ -15,6 +15,7 @@ import (
 	collatorprotocolmessages "github.com/ChainSafe/gossamer/dot/parachain/collator-protocol/messages"
 	parachain "github.com/ChainSafe/gossamer/dot/parachain/runtime"
 	parachaintypes "github.com/ChainSafe/gossamer/dot/parachain/types"
+	"github.com/ChainSafe/gossamer/dot/parachain/util"
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/lib/common"
@@ -137,7 +138,7 @@ func (o *Overseer) processMessages() {
 
 				subsystem = o.nameToSubsystem[parachaintypes.AvailabilityStore]
 
-			case chainapi.ChainAPIMessage[chainapi.Ancestors], chainapi.ChainAPIMessage[chainapi.BlockHeader]:
+			case chainapi.ChainAPIMessage[util.Ancestors], chainapi.ChainAPIMessage[chainapi.BlockHeader]:
 				subsystem = o.nameToSubsystem[parachaintypes.ChainAPI]
 
 			case parachain.RuntimeAPIMessage:
